refactor: buildRoleByCategory を共有ユーティリティに移動
departments/utils/ から roles/utils/ に移動し、
authentication カテゴリからの相対パス参照を解消。
