+++ src/domain/types/roles/implementations/categories/departments/implementations/ia.ts
@@ -1,6 +1,6 @@
 import type Role from "../../../../../role";
+import buildRoleByCategory from "../../../../utils/buildRoleByCategory";
 import { DepartmentRoleCategory } from "../departmentRoleCategory";
-import buildRoleByCategory from "../utils/buildRoleByCategory";
 
 const iaRole: Role = buildRoleByCategory({
   roleCategory: DepartmentRoleCategory,
