+++ src/domain/types/roles/utils/buildRoleByCategory.ts
@@ -1,6 +1,6 @@
-import type Role from "../../../../../role";
-import type RoleCategory from "../../../../../roleCategory";
-import { BOT_PREFIX } from "../../../../constants";
+import type Role from "../../role";
+import type RoleCategory from "../../roleCategory";
+import { BOT_PREFIX } from "../constants";
 
 interface BuildRoleByCategoryProps {
   roleCategory: RoleCategory;
